src/requirements.txt
discord.py
aiohttp
numpy
dynaconf
